journalctl: allow --boot/--list-boots with --file/--dir
This was forbidden in 596a23293d28 ('journalctl: Improve boot ID lookup'),
because the new code does not supported interleaved boots.
But there is no reason for a set of journal files specified with
--dir to be more or less interleaved than a set of files in the
standard location. And with --file, i.e. with a single file, it
is even less likely.

Without this it is hard to diagnose bugs like
https://bugzilla.redhat.com/show_bug.cgi?id=1294002.