shared/install: add information about refusal to operate on linked file to changes
If we don't put the error in the changes array, we get a generic error.
For example, when enabling linked files, "Too many levels of symbolic links".
To get the custom error message, we need to put the error and the path in
*changes.

Before:
Failed to execute operation: Too many levels of symbolic links
Now:
Failed to enable unit, refusing to operate on linked unit file /etc/systemd/system/cat.service

A nicer error message was introduced in 8e20adcaa, but this got broken
again in af3d811352.

Makes #3010 a bit less problematic.