basic/exec-util: add support for synchronous (ordered) execution
After the preparation in previous commits, the change in functionality
is simple. For coding convenience, alarm is prepared *before* any children
are executed, and not before. This shouldn't matter usually, since just
forking of the children should be pretty quick. One could also argue that
this is more correct, because we will also catch the case when (for whatever
reason), forking itself is slow.

The function name is retained, since asynchronous execution will remain
to be the common case.