pid1: attempt to freeze/thaw filesystems that cannot be unmounted/remounted-ro
The advantage is that freezing works even for filesystems which are kept
open by some process or loop device or such. Of course something keeping us
from remounting / read-only is a bug somewhere, but doing freeze/thaw increases
the chances that the fs will be clean after reboot. It seems that XFS is particularly
susceptible to loosing important information.

https://bugzilla.redhat.com/show_bug.cgi?id=1227736