basic: override multi-call argv[0] with $SYSTEMD_INVOKED_AS
This allows us to select multi-call binary behaviour without creating
a symlink. Useful for testing or for cases where we can't rely on the
symlink being in place:

$ build/systemd-fstab-generator
This program takes one or three arguments.
$ SYSTEMD_INVOKED_AS=systemd-sysroot-fstab-check build/systemd-fstab-generator
This program is only useful in the initrd.

In unit files we can use ExecStart=@…, but that is not convenient for
command-line invocations.